Extract signPhotoUrl to shared lib/s3-utils.ts
Consolidates duplicated signPhotoUrl function from profile.ts and
messages.ts into a shared utility. Uses the version with error handling
for graceful degradation when URL signing fails.
