+++ backend/lambdas/api/src/lib/index.ts
@@ -8,3 +8,4 @@ export * from './validation'
 export * from './logger'
 export * from './rate-limit'
 export * from './user'
+export * from './s3-utils'
